Move GraphQLProvider into containers dir
